+++ app/Services/Export/Templates/ModernTemplate.php
@@ -145,13 +145,21 @@ class ModernTemplate implements ExportTemplate
             margin: 4.5em 0 1.5em;
             color: #111111;
         }
-        .toc-entry {
-            margin: 0.3em 0;
-            text-indent: 0;
+        div.mpdf_toc {
+            font-family: {$bodyFontFamily};
+            line-height: {$lineHeight};
         }
-        .toc-entry a {
-            text-decoration: none;
-            color: inherit;
+        span.mpdf_toc_t_level_0 {
+            font-family: {$bodyFontFamily};
+            font-size: 1em;
+            font-weight: normal;
+            color: #333333;
+        }
+        span.mpdf_toc_p_level_0 {
+            font-family: {$bodyFontFamily};
+            font-size: 1em;
+            font-weight: normal;
+            color: #aaaaaa;
         }
         .matter-title {
             font-size: 0.85em;